Ransomware
This project is a basic ransomware simulation designed for educational and security research purposes. It encrypts files and requires a decryption key to unlock them. The ransom.py file generates an encryption key and encrypts specified files. The ransomdecrypter.py file allows users to input the key to decrypt the files.